[UPDATE] Add 2020 subfolder
This is to put my exercise answers in there. This was I can have two notebooks open in separate tabs side-by-side to reference the code.
This is not exactly ideal and can actually be done in Jupyter Lab.

I should consider running this on jupyterLab actually. However there is the issue of deleting answers.