[UPDATE] Complete chapter 3 Kaggle dataset
General comments. This was a fun exercise got a good feel for pipelines.
Also got a good feel for how to go end to end with a project, and how to
improve performance. Spent a good amount of time on data preprocessing
using pipelines which was fun. Also got to play with tabular data which
was also fun.

Something interesting to try would be to add a scaler in the pipeline
and see if that improves performance.